Rename editors and move editors into own folder
