+++ src/pages/platform/editors/steemgig.vue
