+++ src/pages/platform/editors/surpassinggoogle.vue
